+++ packages/types-codec/src/base/Null.ts
@@ -70,6 +70,13 @@ export class Null implements Codec {
   }
 
   /**
+   * @description Converts the value in a best-fit primitive form
+   */
+  public toPrimitive (): null {
+    return null;
+  }
+
+  /**
    * @description Returns the base runtime type name for this instance
    */
   public toRawType (): string {