+++ packages/types-codec/src/native/Float.ts
@@ -113,6 +113,13 @@ export class Float extends Number implements IFloat {
   }
 
   /**
+   * @description Converts the value in a best-fit primitive form
+   */
+  public toPrimitive (): number {
+    return this.toNumber();
+  }
+
+  /**
    * @description Returns the base runtime type name for this instance
    */
   public toRawType (): string {